Break applications out into seperate modules.
