+++ source/fab/builder.py
@@ -5,12 +5,10 @@
 ##############################################################################
 from collections import defaultdict
 from pathlib import Path
-import sys
-from typing import Dict, List, Sequence, Type, Union
+from typing import Dict, List, Type, Union
 
 from fab import FabException
 from fab.database import SqliteStateDatabase, FileInfoDatabase
-from fab.explorer import ExplorerWindow
 from fab.reader import FileTextReader
 from fab.tasks import \
     Task, \
@@ -187,46 +185,3 @@ class Fab(object):
         self._queue.add_to_queue(linker)
         self._queue.check_queue_done()
         self._queue.shutdown()
-
-
-class Grab(object):
-    def __init__(self, workspace: Path):
-        self._workspace = workspace
-
-    def run(self, repositories: Sequence[str]) -> None:
-        pass
-
-
-class Dump(object):
-    def __init__(self, workspace: Path):
-        self._workspace = workspace
-        self._state = SqliteStateDatabase(workspace)
-
-    def run(self, stream=sys.stdout):
-        file_view = FileInfoDatabase(self._state)
-        print("File View", file=stream)
-        for file_info in file_view:
-            print(f"  File   : {file_info.filename}", file=stream)
-            # Where files are generated in the working directory
-            # by third party tools, we cannot guarantee the hashes
-            if file_info.filename.match(f'{self._workspace}/*'):
-                print('    Hash : --hidden-- (generated file)')
-            else:
-                print(f"    Hash : {file_info.adler32}", file=stream)
-
-        fortran_view = FortranWorkingState(self._state)
-        print("Fortran View", file=stream)
-        for info in fortran_view:
-            print(f"  Program unit    : {info.unit.name}", file=stream)
-            print(f"    Found in      : {info.unit.found_in}", file=stream)
-            print(f"    Prerequisites : {', '.join(info.depends_on)}",
-                  file=stream)
-
-
-class Explorer:
-    def __init__(self, workspace: Path):
-        self._state = SqliteStateDatabase(workspace)
-        self._window = ExplorerWindow(self._state)
-
-    def run(self):
-        self._window.mainloop()
