+++ unit-tests/builder_test.py
@@ -1,0 +1,10 @@
+##############################################################################
+# (c) Crown copyright Met Office. All rights reserved.
+# For further details please refer to the file COPYRIGHT
+# which you should have received as part of this distribution
+##############################################################################
+# Each class in builder.py embodies an entire utility. In effect the
+# "unit" is the entire tool. As such any unit testing would effectively be
+# system testing.
+#
+# TODO: We may wish to consider how much stuff is in these top level objects.
